+++ lib/rspec/matchers/built_in/eq.rb
@@ -1,3 +1,7 @@
+# encoding: UTF-8
+
+require 'date'
+
 module RSpec
   module Matchers
     module BuiltIn
@@ -7,14 +11,30 @@ module RSpec
         end
 
         def failure_message_for_should
-          "\nexpected: #{expected.inspect}\n     got: #{actual.inspect}\n\n(compared using ==)\n"
+          "\nexpected: #{format_object(expected)}\n     got: #{format_object(actual)}\n\n(compared using ==)\n"
         end
 
         def failure_message_for_should_not
-          "\nexpected: value != #{expected.inspect}\n     got: #{actual.inspect}\n\n(compared using ==)\n"
+          "\nexpected: value != #{format_object(expected)}\n     got: #{format_object(actual)}\n\n(compared using ==)\n"
         end
 
         def diffable?; true; end
+
+      private
+
+        def format_object(object)
+          case object
+          when Time
+            microseconds = ("%10.6f" % object.to_f)[-6, 6]
+            offset = "#{object.gmt_offset < 0 ? "-" : "+"}%02d%02d" % (object.gmt_offset/60).abs.divmod(60)
+            object.strftime("%Y-%m-%d %H:%M:%S #{offset} (+ #{microseconds} μs)")
+          when DateTime
+            microseconds = object.strftime("%6N")[0,6]
+            object.strftime("%Y-%m-%d %H:%M:%S %z (+ #{microseconds} μs)")
+          else
+            object.inspect
+          end
+        end
       end
     end
   end